Testing different options for voting functionality
The `testingVotingFunctionality` function can now be used to test different ideas about how users vote i.e. we can perhaps just assign the amount of tokens which are staked for that user against an address => uint256 mapping for a voting option. The address => uint256 with the highest value at the end wins. This is a natural way to weight options without performing many calculations.